service/src/main/java/uk/gov/hmcts/reform/fpl/components/ChildSolicitorPolicyConverter.java
package uk.gov.hmcts.reform.fpl.components;

import org.springframework.stereotype.Component;
import uk.gov.hmcts.reform.ccd.model.Organisation;
import uk.gov.hmcts.reform.ccd.model.OrganisationPolicy;
import uk.gov.hmcts.reform.fpl.enums.SolicitorRole;
import uk.gov.hmcts.reform.fpl.model.Child;
import uk.gov.hmcts.reform.fpl.model.common.Element;

import java.util.Optional;

import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;

@Component
public class ChildSolicitorPolicyConverter {
    public OrganisationPolicy generate(SolicitorRole solicitorRole,
                                       Optional<Element<Child>> optionalRespondentElement) {
        return OrganisationPolicy.builder()
            .organisation(getOrganisation(optionalRespondentElement))
            .orgPolicyCaseAssignedRole(solicitorRole.getCaseRoleLabel())
            .build();
    }

    private Organisation getOrganisation(Optional<Element<Child>> optionalRespondentElement) {
        return optionalRespondentElement.map(Element::getValue)
            .filter(child ->
                isNotEmpty(child.getRepresentative()) && isNotEmpty(child.getRepresentative().getOrganisation()))
            .map(child -> child.getRepresentative().getOrganisation())
            .orElse(Organisation.builder().build());
    }

}
